requirements.txt
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.0.11
enum34==1.1.6
functools32==3.2.3.post2
ipython==5.4.1
ipython-genutils==0.2.0
matplotlib==2.0.2
numpy==1.13.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.7
tifffile==0.12.1
traitlets==4.3.2
wcwidth==0.1.7
Pillow==4.2.0
